split main pages js files
