+++ js/script.js
@@ -1,4 +1,5 @@
 require('./main-page/validator.js');
-require('./main-page/animation.js');
+require('./main-page/mask.js');
+require('./main-page/mobile-menu.js');
 require('./slider/slider-images.js');
 require('./slider/slider-pagination.js');
